refactor: 키보드 스크롤 액션 함수 유틸파일에 빼둠
